Adding support for compleTypes
This is just an initial checkin for handling complex types
It is mostlikely buggy and has been sloppily thrown in
without proper testing.  This will have to change but, since
schemavore is a hobby I will afford myself this slip because
the urge to add this and get over the holiday slup had overcame
me.